Implemented class initialProcess to handle the "(" and ")" operators when read.
In class "int eval" we clear the operator and operand stack while pushing the null operator into the stack.

Inner if statement finds the token for the parenthesis operator and implements a push depending on the operator.
Created the newOperator to hold our Operator token, if the stack is empty then it fills with the newOperator.

The last while loop value changed to -1 because the lowest priority value created is -1 for the HashOperator().